Scripts-SS/requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.2.0
cffi==1.14.5
colorama==0.4.3
cryptography==3.4.7
html-testRunner==1.2.1
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.1
paramiko==2.7.2
pluggy==0.13.1
py==1.8.1
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.6
pytest==5.3.5
pytest-html==2.0.1
pytest-metadata==1.8.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
